chore: 동아리 목록을 보여주는 컴포넌트를 공통화
