+++ core/ui/src/main/res/values/strings.xml
@@ -17,4 +17,5 @@
     <!-- Club -->
     <string name="club_card_tag_recruitment_complete">모집종료</string>
     <string name="club_card_tag_d_day">D-%s</string>
+    <string name="club_list_no_item">등록된 동아리가 없어요!</string>
 </resources>
